forensic/dashboards/requirements.txt
# Forensic Dashboard Requirements
# ================================

# Core web framework
streamlit>=1.28.0
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0

# Machine learning and metrics
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
altair>=4.2.0

# SHAP analysis
shap>=0.42.0

# Evidently integration
evidently>=0.4.0

# Authentication and security
streamlit-authenticator>=0.2.1
cryptography>=3.4.8
bcrypt>=3.2.0
PyJWT>=2.4.0

# Database and storage
sqlite3
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Export capabilities
reportlab>=3.6.0
openpyxl>=3.0.9
xlswriter>=3.0.0
fpdf2>=2.5.0

# Real-time monitoring
schedule>=1.2.0
watchdog>=2.1.9

# API and HTTP clients
requests>=2.28.0
urllib3>=1.26.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
pathlib>=1.0.1
jsonschema>=4.17.0
pydantic>=1.10.0

# Logging and monitoring
loguru>=0.6.0
structlog>=22.1.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0